feat: write basic Hacker News client
After a little trial and error, I decided to write my own simple Hacker News client that operates primarily based on an async iterator.

My first attempt was using the FireBase client library, since Hacker News provides the API through FireBase (and they recommend using the client). However, I was turned off by the complexity of the library and the fact that it provided a lot of functionality that I wouldn't end up actually needing. In reality I just need to make a few API calls, and doing that manually would be easy enough.

When designing the client library, I first attempted to base it on an Observable, thinking that it would allow me to easily model the asynchronous delivery of each of the items. However, the direction of control was wrong; I wanted the UI to determine when it asked for each new item and Observables are simple even emitters.

An async generator ended up being the right abstraction for providing each item from the Hacker News API; the need to make the initial API call for the list is hidden from the user of the iterator, and the consumer is able to determine when new items are requested. This will become the means by which the application can load more Items from the API as they scroll.