feat: basic loading of Items on scroll
In order to achieve the desired UX around loading posts one at a time as the user scrolls, we can build a React component that leverages an IntersectionObserver to determine when posts need to be loaded.

By creating a list with a loading indicator at the bottom, we can create a basic logic flow that looks something like

    while (loadingIndicatorIsVisible) {
      loadAdditionalPost()
    }

At the point that the view is "full" and the loading indicator is no longer on screen, we stop reading elements from the iterator.

If the user scrolls the list and the loading indicator comes into view again, we start loading posts from the API again until loading indicator is no longer visible.